Take out IndexationStep component from ConfigurationWorkout
